修正

+++ problems/1word-mod/runtime/1word-31/algos/barrett_ceil_64.hpp
@@ -3,7 +3,7 @@
 struct DIV {
  u32 b;
  constexpr DIV(u32 b): b(b), x(u64(-1) / b + 1) {}
- constexpr u32 mod(u32 a) const { return a - u64(u32(a * x >> 64)) * b; }
+ constexpr u32 mod(u32 a) const { return a - u64(u32(u128(a) * x >> 64)) * b; }
 private:
  u64 x;
 };
